requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
prettytable==0.7.2
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
tqdm==4.43.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
